move provenance function to a separate module
